fixed gdumb

+++ memory.py
@@ -70,8 +70,7 @@ class ClassificationMemoryManager(Dataset):
                 x_i for label in labels for x_i in self.data[label]
             ]
             labels = [
-                label.to(torch.uint8)
-                for label in labels for _ in self.data[label]
+                label for label in labels for _ in self.data[label]
             ]
         else:
             data = []
